# openapi/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: oojuniin@outlook.com
    name: Junior Lima
  description: "Contrato REST do budget-service. Este documento é a fonte da verdade\
    \ para qualquer client (frontend, agentes de IA) que consuma esta API — não assuma\
    \ comportamento não descrito aqui."
  title: Budget Service API
  version: 0.0.1
servers:
- description: Relativo ao host onde a API estiver publicada
  url: /
security:
- bearer-jwt: []
paths:
...
  /api/v1/revenue-types:
    get:
      operationId: findAll_1
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueTypeDTO"
          description: OK
      tags:
      - revenue-type-controller
    post:
      operationId: save_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueTypeDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueTypeDTO"
          description: Created
      tags:
      - revenue-type-controller
  /api/v1/revenue-types/{id}:
    delete:
      operationId: delete_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - revenue-type-controller
    get:
      operationId: findById_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueTypeDTO"
          description: OK
      tags:
      - revenue-type-controller
    put:
      operationId: update_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueTypeDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueTypeDTO"
          description: OK
      tags:
      - revenue-type-controller
  /api/v1/revenue-types/{id}/history:
    get:
      operationId: history_3
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueTypeRevisionDTO"
          description: OK
      tags:
      - revenue-type-controller
...
  /api/v1/revenues/{id}:
    delete:
      operationId: delete_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - revenue-controller
    get:
      operationId: findById_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueDTO"
          description: OK
      tags:
      - revenue-controller
    put:
      operationId: update_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RevenueDTO"
          description: OK
      tags:
      - revenue-controller
  /api/v1/revenues/{id}/history:
    get:
      operationId: history_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueRevisionDTO"
          description: OK
      tags:
      - revenue-controller
...
  /api/v1/spending-types/{id}:
    delete:
      operationId: delete_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - spending-type-controller
    get:
      operationId: findById_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingTypeDTO"
          description: OK
      tags:
      - spending-type-controller
    put:
      operationId: update_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendingTypeDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingTypeDTO"
          description: OK
      tags:
      - spending-type-controller
  /api/v1/spending-types/{id}/history:
    get:
      operationId: history_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingTypeRevisionDTO"
          description: OK
      tags:
      - spending-type-controller
...
  /api/v1/spendings/{id}:
    delete:
      operationId: delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - spending-controller
    get:
      operationId: findById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingDTO"
          description: OK
      tags:
      - spending-controller
    put:
      operationId: update
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendingDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpendingDTO"
          description: OK
      tags:
      - spending-controller
  /api/v1/spendings/{id}/history:
    get:
      operationId: history
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingRevisionDTO"
          description: OK
      tags:
      - spending-controller
components:
  schemas:
...
    RevenueDTO:
      type: object
      properties:
        date:
          type: string
          format: date
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        typeId:
          type: string
          format: uuid
        typeName:
          type: string
        value:
          type: number
      required:
      - date
      - typeId
      - value
    RevenueRevisionDTO:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        date:
          type: string
          format: date
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
        typeId:
          type: string
          format: uuid
        value:
          type: number
    RevenueTypeDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        includeInMonthlyTotals:
          type: boolean
        includeInTotals:
          type: boolean
        name:
          type: string
          minLength: 1
      required:
      - name
    RevenueTypeRevisionDTO:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        id:
          type: string
          format: uuid
        includeInMonthlyTotals:
          type: boolean
        includeInTotals:
          type: boolean
        name:
          type: string
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
...
    SpendingDTO:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
          maxLength: 250
          minLength: 3
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        typeId:
          type: string
          format: uuid
        typeName:
          type: string
        value:
          type: number
        wasPaid:
          type: boolean
      required:
      - date
      - typeId
      - value
      - wasPaid
    SpendingRevisionDTO:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        id:
          type: string
          format: uuid
        referenceDate:
          type: string
          format: date
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
        typeId:
          type: string
          format: uuid
        value:
          type: number
        wasPaid:
          type: boolean
    SpendingTypeDTO:
      type: object
      properties:
        category:
          type: string
          enum:
          - ESSENTIAL
          - PERSONAL
          - SAVINGS
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
      required:
      - category
      - name
    SpendingTypeRevisionDTO:
      type: object
      properties:
        category:
          type: string
          enum:
          - ESSENTIAL
          - PERSONAL
          - SAVINGS
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        revision:
          type: integer
          format: int32
        revisionType:
          type: string
...
  securitySchemes:
    bearer-jwt:
      bearerFormat: JWT
      name: bearer-jwt
      scheme: bearer
      type: http